big_data/requirements.txt
requests
Producer
fastavro
BytesIO
